Fixed two wolf bugs, puppy mills are active!
Basically, wolves would eat an endless amount of food when breedable. Puppies would not eat because the code for baby animals skips right to starving for some reason. Strangely, baby cows and things would eat, but not puppies. Possible that baby herbivores override the hunger update in a unique way.